Now CAMERA WORKS!

+++ ImportNodesFromMaya/ImportNodesFromMaya/PluginDeclarations.h
@@ -23,7 +23,7 @@ struct CameraHeader
 	bool isPerspective;
 	float projectionMatrix[16];
 	float translation[3];
-	float rotation[4];
+	double rotation[4];
 	float nearPlane;
 	float farPlane;
 	float fieldOfView;
